feat: remove service completely
